feat: added shipping screen
