+++ lib/services/auth_service.dart
@@ -3,7 +3,7 @@ import 'package:http/http.dart' as http;
 
 class AuthService {
   static Future<Map<String, dynamic>> login(String email, String password) async {
-    final url = Uri.parse('http://antarticdonkeys.com/login?email=$email&password=$password');
+    final url = Uri.parse('https://api-akira.antarticdonkeys.com/login?email=$email&password=$password');
 
     try {
       final response = await http.get(url);
@@ -19,7 +19,7 @@ class AuthService {
   }
 
   static Future<Map<String, dynamic>> getUserData(int userId) async {
-    final response = await http.get(Uri.parse('http://antarticdonkeys.com/users/$userId'));
+    final response = await http.get(Uri.parse('https://api-akira.antarticdonkeys.com/users/$userId'));
 
     if (response.statusCode == 200 ) {
       return jsonDecode(response.body);
@@ -28,7 +28,7 @@ class AuthService {
     }
   }
   static Future<Map<String, dynamic>> register(String name, String surname, String email, String password, String numberCellphone, String payment) async {
-    final url = Uri.parse('http://antarticdonkeys.com/register');
+    final url = Uri.parse('https://api-akira.antarticdonkeys.com/register');
 
     final response = await http.post(
       url,
